Cleaned and renamed SPI file
